in production

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: imiqor <imiqor@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/22 19:20:35 by macbookair        #+#    #+#             */
-/*   Updated: 2024/11/15 00:36:08 by imiqor           ###   ########.fr       */
+/*   Updated: 2025/03/09 02:50:01 by imiqor           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i] != 0)
